tests/wpt/tests/infrastructure/expected-fail/window-onload-test.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
window.onload = () => {
  test((t) => {
    assert_true(true,'This should pass');
  }, 'test 1');
  test((t) => {
    assert_true(false,'This should FAIL');
  }, 'test 2');
  test((t) => {
    assert_true(false,'This should FAIL');
  }, 'test 3');

  promise_test(async t => {
    assert_true(false,'This should FAIL');
  },'promise 1');
  promise_test(async t => {
    assert_true(false,'This should FAIL');
  },'promise 2');
  promise_test(async t => {
    assert_true(false,'This should FAIL');
  },'promise 3');
};
</script>
