third_party/blink/web_tests/external/wpt/infrastructure/expected-fail/window-onload-test-expected.txt

This is a testharness.js-based test.
[FAIL] test 2
  assert_true: This should FAIL expected true got false
[FAIL] test 3
  assert_true: This should FAIL expected true got false
[FAIL] promise 1
  assert_true: This should FAIL expected true got false
[FAIL] promise 2
  assert_true: This should FAIL expected true got false
[FAIL] promise 3
  assert_true: This should FAIL expected true got false
Harness: the test ran to completion.
